DiscordBot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cffi==1.14.6
chardet==4.0.0
click==7.1.2
discord.py==1.7.3
distlib==0.3.2
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.0.12
idna==3.2
keyboard==0.13.5
MouseInfo==0.1.3
multidict==5.1.0
numpy==1.21.2
openpyxl==3.0.7
pafy==0.5.5
pandas==1.3.2
platformdirs==2.3.0
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.4.0
pynput==1.7.5
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.0
pytweening==1.0.4
pytz==2021.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
six==1.16.0
typing-extensions==3.10.0.2
virtualenv==20.7.2
xlrd==2.0.1
XlsxWriter==3.0.1
yarl==1.6.3
youtube-dl==2021.6.6
